Moved Uniplayer into Controls
