+++ src/modules/controls/components/Duration.jsx
