+++ src/modules/controls/components/PlayerBox/Player.jsx
